+++ src/hoist/__init__.py
@@ -3,4 +3,6 @@ from .client import Connection
 from .exceptions import *
 from .message import Message
 from .server import Server
-from .utils import connect, connect_to, debug, main, serve, start
+from .utils import (
+    connect, connect_no_ctx, connect_with, debug, main, serve, start
+)