+++ src/hoist/_typing.py
@@ -1,6 +1,6 @@
 from typing import (
     TYPE_CHECKING, Any, Awaitable, Callable, Dict, List, Optional, Tuple, Type,
-    TypeVar, Union
+    Union
 )
 
 from typing_extensions import Protocol
@@ -13,9 +13,6 @@ if TYPE_CHECKING:
     from ._operations import OperatorParam
     from .message import Message
     from .server import Server
-
-
-_T = TypeVar("_T")
 
 
 class DataclassLike(Protocol):
@@ -39,20 +36,20 @@ _PyBuiltins = Union[str, float, int, bool, dict, None]
 JSONLike = Union[_PyBuiltins, HasDict]
 Payload = Dict[str, Any]
 Operator = Union[
-    Callable[[_T], Awaitable[JSONLike]],
+    Callable[[Any], Awaitable[JSONLike]],
     Callable[[], Awaitable[JSONLike]],
-    Callable[["Server", _T], Awaitable[JSONLike]],
+    Callable[["Server", Any], Awaitable[JSONLike]],
     Callable[["Server"], Awaitable[JSONLike]],
 ]
 SchemaNeededType = Union[Type[Any], Tuple[Optional[Type[Any]], ...]]
 Schema = Dict[str, SchemaNeededType]
-OperationData = Tuple[Operator[_T], "OperatorParam", bool]
-Operations = Dict[str, OperationData[_T]]
+OperationData = Tuple[Operator, "OperatorParam", bool]
+Operations = Dict[str, OperationData]
 UrlLike = Union[str, "URL"]
 LoginFunc = Callable[["Server", str], Awaitable[bool]]
 ResponseErrors = Dict[int, Tuple[str, str]]
 Listener = Union[
-    Callable[["Message", _T], Awaitable[None]],
+    Callable[["Message", Any], Awaitable[None]],
     Callable[["Message"], Awaitable[None]],
     Callable[[], Awaitable[None]],
 ]