+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 -   Implemented message receiving with replies.
 -   Added fancy mode.
 -   Added dynamic arguments on message listeners.
+-   Added support for custom operations.
+-   Changed `connect_to` to `connect_with`.
+-   Improved internal API.
 
 ## `1.0.0-alpha.2`